+++ src/ReverseProxy/Service/Proxy/HttpProxy.cs
@@ -379,11 +379,6 @@ namespace Microsoft.ReverseProxy.Service.Proxy
         {
             // Transforms that were run in the first pass.
             HashSet<string> transformsRun = null;
-
-            // HttpClient wrongly uses comma (",") instead of semi-colon (";") as a separator for Cookie headers.
-            // To mitigate this, we collect them, concatenate them manually and put them back as a single header value.
-            List<string> cookies = null;
-
             if (transforms.CopyRequestHeaders ?? true)
             {
                 foreach (var header in context.Request.Headers)
@@ -411,10 +406,7 @@ namespace Microsoft.ReverseProxy.Service.Proxy
                         }
                     }
 
-                    if (!CatchCookies(headerName, headerValue, ref cookies))
-                    {
-                        AddHeader(destination, headerName, headerValue);
-                    }
+                    AddHeader(destination, headerName, headerValue);
                 }
             }
 
@@ -427,40 +419,9 @@ namespace Microsoft.ReverseProxy.Service.Proxy
                     headerValue = transform.Apply(context, headerValue);
                     if (!StringValues.IsNullOrEmpty(headerValue))
                     {
-                        if (!CatchCookies(headerName, headerValue, ref cookies))
-                        {
-                            AddHeader(destination, headerName, headerValue);
-                        }
+                        AddHeader(destination, headerName, headerValue);
                     }
                 }
-            }
-
-            if (cookies != null && cookies.Count > 0)
-            {
-                AddHeader(destination, "Cookie", String.Join("; ", cookies));
-            }
-
-            // Returns true when a cookie header is collected and therefore shouldn't be added to the destination yet.
-            static bool CatchCookies(string headerName, StringValues value, ref List<string> cookies)
-            {
-                if (headerName != "Cookie")
-                {
-                    return false;
-                }
-
-                cookies ??= new List<string>();
-                if (value.Count == 1)
-                {
-                    string headerValue = value;
-                    cookies.Add(headerValue);
-                }
-                else
-                {
-                    string[] headerValues = value;
-                    cookies.AddRange(headerValues);
-                }
-
-                return true;
             }
 
             // Note: HttpClient.SendAsync will end up sending the union of
@@ -469,6 +430,13 @@ namespace Microsoft.ReverseProxy.Service.Proxy
             // as long as they appear in one (and only one, otherwise they would be duplicated).
             static void AddHeader(HttpRequestMessage request, string headerName, StringValues value)
             {
+                // HttpClient wrongly uses comma (",") instead of semi-colon (";") as a separator for Cookie headers.
+                // To mitigate this, we concatenate them manually and put them back as a single header value.            
+                if (headerName == HeaderNames.Cookie && value.Count > 1)
+                {
+                    value = String.Join("; ", value);
+                }
+
                 if (value.Count == 1)
                 {
                     string headerValue = value;
